4032 - useActions: Separate hook with return types
